// src/types.ts
export interface FormValues {
  sphere: string;
  profit: number;
  budget: number;
  budgetType: string;
  loanAmount: string;
  loanTerm: string;
  interestRate: string;
  monthlyPayment: number;
  legalForm: string;
  taxRate: string;
  expenses: Array<any>;
  incomes: Array<any>;
  employees: Array<any>;
}